face_morphing/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
numpy==1.18.2
Pillow==7.1.1
requests==2.23.0
urllib3==1.25.9
